chore: move function from types.ts to utils.ts
